+++ NEWS.md
@@ -188,6 +188,13 @@ Theme fixes:
 
 ## v0.6.1
 
+### New features
+
+- New dedicated `tinyplot()` method for `ts` time series. Internally, this sets
+  up a long data frame with columns `Time`, `Value`, and `Series` and then calls
+  the formula method with different possible specifications for the `by` and `facet`
+  variables. (#558 @zeileis)
+
 ### Aesthetic changes
 
 - The legend plot characters for the `"pointrange"` and `"errorbar"` types now